+++ kinto_signer/updater.py
@@ -181,13 +181,13 @@ class LocalUpdater(object):
                                   record=created,
                                   action=ACTIONS.CREATE)
 
-        # Set the permissions on the destination collection.
-        # With the current implementation, the destination is not writable by
-        # anyone and readable by everyone.
-        # https://github.com/Kinto/kinto-signer/issues/55
-        permissions = {'read': ("system.Everyone",)}
-        self.permission.replace_object_permissions(
-            self.destination_collection_uri, permissions)
+            # Set the permissions on the destination collection.
+            # With the current implementation, the destination is not writable by
+            # anyone and readable by everyone.
+            # https://github.com/Kinto/kinto-signer/issues/55
+            permissions = {'read': ("system.Everyone",)}
+            self.permission.replace_object_permissions(
+                self.destination_collection_uri, permissions)
 
     def _get_records(self, rc, last_modified=None):
         # If last_modified was specified, only retrieve items since then.